New update

+++ philo/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: abablil <abablil@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/18 21:14:58 by abablil           #+#    #+#             */
-/*   Updated: 2024/05/16 15:14:43 by abablil          ###   ########.fr       */
+/*   Updated: 2024/05/16 21:59:01 by abablil          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ typedef struct s_philo
 void		delete_data(t_data *data, int free_data);
 
 // Utils
-int			get_number(char *str);
+long		get_number(char *str);
 int			exit_program(char *message, t_data *data, int free_data);
 void		end_simulation(t_data *data);
 void		print(t_philo *philo, int message);
