+++ philo_bonus/utils.c
@@ -6,7 +6,7 @@
 /*   By: abablil <abablil@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/20 19:59:53 by abablil           #+#    #+#             */
-/*   Updated: 2024/05/16 15:08:47 by abablil          ###   ########.fr       */
+/*   Updated: 2024/05/16 21:59:35 by abablil          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,11 +49,11 @@ void	exit_program(char *message, int free, t_data *data, int should_kill)
 	exit(1);
 }
 
-int	get_number(char *str)
+long	get_number(char *str)
 {
-	int	i;
-	int	result;
-	int	sign;
+	int		i;
+	long	result;
+	int		sign;
 
 	i = 0;
 	result = 0;
